requirements.txt
aiogram>=3.0,<4.0
python-dotenv>=1.0,<2.0
sqlalchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0